Add a verbose column to indicate that DWM has been notified of a particular present. This is mainly relevant for independent flip, so on-screen display of this fact is only done for these modes.
DWM may request to be notified of an app's presents if it has need of showing them somewhere besides the single screen where the app has entered independent flip - for example, opening a thumbnail view of the window on a second monitor.

This can be a root cause of some unexpected performance impact while in independent flip, but on its own is not an interesting performance metric, which is why it is under the verbose mode.