Preliminary support for Win7.
 Interesting things about Win7:

1. DxgKrnl wasn't yet on the "Crimson" manifest-based ETW providers. This means that enabling the provider is done through a different GUID, and events are identified by their own unique GUIDs instead of the provider GUID and an event ID. This required two bits of refactoring:
a. The TraceSession now has two maps - one for providers to enable and the properties used to do so, and a separate one for GUID -> handler mappings. For Crimson events, the handler is expected to understand events by ID, but for classic events, each handler only expects one event type.
b. Most DxgKrnl event handling was split out into common functions which can be invoked by either classic or Crimson event handling.

2. There's no Win32K events for flip model. As a result, I don't expect to be able to handle sync intervals other than 1 or DO_NOT_SEQUENCE presents properly, but at least I can properly classify them. This is probably fine, since it was limited to D3D9Ex and probably only used by Windows Media Player.

3. There's no DxgkPresent event. This makes distinguishing DWM-on blts from DWM-off blts actually somewhat difficult. I think I was able to work around this.

I have a lot of testing to do before this is ready.