Fixes for Win7 support.
Brought back the VistaBlt path, since I forgot that the blt event's window is null for that. However, on Win7, there's no "TokenData" field in SubmitPresentHistoryToken, only a pointer to the token, so I don't think I can correlate it with the DWM events to get a proper window. That means that I have to assume which presents are getting composed (all) and when they're making it to screen (next DWM present).

A few other bugfixes (uninitialized data, missing assignment, missing function calls, bad 'stuck' check).

Added the Win7 DWM provider GUID. It has most of the same events.