Fix bug corrupting token storage in mDxgKrnlPresentHistoryTokens
There was an issue where a present gets handed off to a DWM (becoming a
mDependentPresent) and is eventually completed when the DWM present
completes.

The token used by that dependent present can be reused for a new
present.  If this happens, and the new present is live in
mDxgKrnlPresentHistoryTokens when the dependent present is completed,
then CompletePresent() did erase the mDxgKrnlPresentHistoryTokens entry
causing the present to be lost and eventually labeled as dropped.