Change the default behaviour to not stop existing sessions.
It can be useful to stop existing sessions (e.g., if a previous
PresentMon process failed to shut down and stop the session
successfully).  However, the default should be (and now is) an error
when starting PresentMon if a same-named session is already active.
This prevents breaking traces that are already running.

Use the new -stop_existing_session command line argument to use the
previous behavior.